Move all the product methods to a trait to keep the Client class clean
